Test proj ----modified 2!

+++ proj1b/OffByOne.java
@@ -2,6 +2,6 @@ public class OffByOne implements CharacterComparator {
     @Override
     public boolean equalChars(char a, char b) {
         //return 1 == (a - b) || -1 == (a - b);
-        return (Math.abs(a - b) == 1);
+        return (Math.abs((int)a - (int)b) == 1);
     }
 }
